feat: make instance requests scoped in order to support scoped role such as online maintainer
